OEL-100: Move 2 tests under src/Api.
